Changed way how complex types are stored
First of all GSettings arrays are now stored as array in Xfconf.

Second, the way how complex or non-convertible types are stored has changed. Previously the GVariant was serialized and just stored as string in Xfconf. But this was problem that the signature was unknown, e.g. when editting the value. Now these GVariants are stored as an array containing three elements. The first element is a magic number ('GVar' as 32 bit unsigned integer), the second element is the signature or the type description of the GVariant stored in the third element - the serialized string representation of the GVariant as it was stored previously. This prevents the edit of these value in xfce4-settings-editor but I expect is to be more stable and reliable.